requirements.txt
click==6.7
eventlet==0.23.0
Flask==1.0.2
greenlet==0.4.13
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
python-engineio==2.1.0
python-socketio==1.9.0
six==1.11.0
Werkzeug==0.14.1
